# docs/swagger.yaml
components:
  schemas:
    models.CellTexture:
      properties:
        createdAt:
          example: 1700000000
          type: integer
        displayName:
          example: Tundra Grass
          maxLength: 64
          minLength: 2
          type: string
        key:
          example: tundra.grass
          maxLength: 64
          minLength: 2
          type: string
        updatedAt:
          example: 1700000000
          type: integer
      required:
      - createdAt
      - displayName
      - key
      - updatedAt
      type: object
    models.NewDiscordSession:
      properties:
        authCode:
          type: string
      required:
      - authCode
      type: object
    models.NewEmailSession:
      properties:
        email:
          example: john.doe@email.com
          maxLength: 255
          minLength: 5
          type: string
        password:
          minLength: 8
          type: string
      required:
      - email
      type: object
    models.NewProfile:
      properties:
        avatarUri:
          example: http://cdn.discord.com/some-image-id
          maxLength: 1024
          type: string
        username:
          example: Krusher99
          maxLength: 64
          minLength: 2
          type: string
      required:
      - username
      type: object
    models.NewUser:
      properties:
        email:
          example: john.doe@email.com
          maxLength: 255
          minLength: 5
          type: string
      required:
      - email
      type: object
    models.Profile:
      properties:
        avatarUri:
          example: http://cdn.discord.com/some-image-id
          maxLength: 1024
          type: string
        createdAt:
          example: 1700000000
          type: integer
        updatedAt:
          example: 1700000000
          type: integer
        userId:
          example: 00000000-0000-0000-0000-000000000000
          maxLength: 36
          minLength: 36
          type: string
        username:
          example: Krusher99
          maxLength: 64
          minLength: 2
          type: string
      required:
      - createdAt
      - updatedAt
      - userId
      - username
      type: object
    models.Provider:
      properties:
        authUrl:
          type: string
        name:
          example: Discord
          type: string
      required:
      - authUrl
      - name
      type: object
    models.User:
      properties:
        email:
          example: john.doe@email.com
          maxLength: 255
          minLength: 5
          type: string
        id:
          example: 00000000-0000-0000-0000-000000000000
          maxLength: 36
          minLength: 36
          type: string
        isAdmin:
          default: false
          type: boolean
      required:
      - email
      - id
      - isAdmin
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
externalDocs:
  description: ""
  url: ""
info:
  description: Web API for OpenDungeon
  title: OpenDungeon API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/cell-textures:
    get:
      description: List all existing cell textures.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.CellTexture'
                type: array
          description: List of cell textures
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: List cell textures
      tags:
      - Cell Textures
    post:
      description: Create a new cell texture.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
              - title: key
                type: string
              - title: displayName
                type: string
              - title: file
                type: file
          multipart/form-data:
            schema:
              type: object
        description: Texture key | Texture display name | 64x64 image file
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.CellTexture'
          description: Newly created texture details
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "415":
          content:
            application/json:
              schema:
                type: string
          description: Unsupported media type
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: Create cell texture
      tags:
      - Cell Textures
  /api/cell-textures/{key}:
    get:
      description: Get an existing cell texture.
      parameters:
      - description: Key
        in: path
        name: key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            image/png:
              schema:
                type: file
          description: Texture content
        "400":
          content:
            image/png:
              schema:
                type: string
          description: Bad request
        "404":
          content:
            image/png:
              schema:
                type: string
          description: Not found
        "500":
          content:
            image/png:
              schema:
                type: string
          description: Server error
      summary: Get cell texture
      tags:
      - Cell Textures
  /api/profiles/me:
    get:
      description: Get the profile for the authenticated user.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Profile'
          description: OK
        "401":
          content:
            application/json:
              schema:
                type: string
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not found
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: Get user's profile
      tags:
      - Profiles
    put:
      description: Create or replace the profile for the authenticated user.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/models.NewProfile'
                description: Profile data
                summary: request
        description: Profile data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Profile'
          description: Created
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                type: string
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: Create or replace user's profile
      tags:
      - Profiles
  /api/providers:
    get:
      description: List available auth providers.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Provider'
                type: array
          description: Available auth providers
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: List providers
      tags:
      - Providers
  /api/sessions:
    delete:
      description: Delete the active session.
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                type: string
          description: Unauthorized
      summary: Delete session
      tags:
      - Sessions
  /api/sessions/discord:
    post:
      description: |-
        Create a new session using Discord authentication.
        If a user with the associated email does not exist, it is created.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/models.NewDiscordSession'
                description: Discord credentials
                summary: request
        description: Discord credentials
        required: true
      responses:
        "201":
          description: Session id cookie
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: Create session with Discord
      tags:
      - Sessions
  /api/sessions/email:
    post:
      description: Create a new session using email authentication.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/models.NewEmailSession'
                description: Email credentials
                summary: request
        description: Email credentials
        required: true
      responses:
        "201":
          description: Session id cookie
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: Create session with email
      tags:
      - Sessions
  /api/users:
    post:
      description: Create a new user.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/models.NewUser'
                description: New user information
                summary: request
        description: New user information
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.User'
          description: Created
        "409":
          content:
            application/json:
              schema:
                type: string
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Server error
      summary: Create user
      tags:
      - Users
servers:
- url: http://localhost:8000/api